minor clean up; moved pigpio tests

+++ libraries/pi4j-library-pigpio/src/main/java/com/pi4j/library/pigpio/impl/PiGpioSocketImpl.java
@@ -119,11 +119,11 @@ public class PiGpioSocketImpl extends PiGpioSocketBase implements PiGpio {
      * @see <a href="http://abyz.me.uk/rpi/pigpio/cif.html#gpioVersion">PIGPIO::gpioVersion</a>
      */
     @Override
-    public long gpioVersion() throws IOException {
+    public int gpioVersion() throws IOException {
         logger.trace("[VERSION] -> GET VERSION");
         validateReady();
         PiGpioPacket result = sendCommand(PIGPV);
-        long version = result.result();
+        int version = result.result();
         logger.trace("[VERSION] <- RESULT={}", version);
         return version;
     }
